Split transfers epics in their own submodule
